# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type FavoriteMovie struct {
	UserID  pgtype.UUID `json:"user_id"`
	MovieID pgtype.UUID `json:"movie_id"`
}

type Movie struct {
	ID        pgtype.UUID      `json:"id"`
	Title     string           `json:"title"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}

type MovieRatingView struct {
	MovieID     pgtype.UUID `json:"movie_id"`
	AmountRates int64       `json:"amount_rates"`
	Rating      float64     `json:"rating"`
}

type RatedMovie struct {
	UserID  pgtype.UUID `json:"user_id"`
	MovieID pgtype.UUID `json:"movie_id"`
	Rating  int16       `json:"rating"`
}

type UserDatum struct {
	ID        pgtype.UUID      `json:"id"`
	Name      string           `json:"name"`
	Login     string           `json:"login"`
	Password  string           `json:"-"`
	IsAdmin   bool             `json:"-"`
	CreatedAt pgtype.Timestamp `json:"created_at"`
}