# docs/swagger.yaml
definitions:
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
  reqmodel.FavoriteMovieListResponse:
    properties:
      favorite_movie_ids:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  reqmodel.MovieListResponse:
    properties:
      movie_list:
        items:
          $ref: '#/definitions/sqlc.Movie'
        type: array
    type: object
  reqmodel.MovieRequest:
    properties:
      title:
        type: string
    type: object
...
  sqlc.FavoriteMovie:
    properties:
      movie_id:
        type: string
      user_id:
        type: string
    type: object
  sqlc.GetMovieByIDRow:
    properties:
      amount_rates:
        type: integer
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      rating:
        type: number
      title:
        type: string
    type: object
...
  sqlc.Movie:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
  /movie:
    get:
      consumes:
      - application/json
      description: Get movie list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.MovieListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show movie list
      tags:
      - movie
    post:
      consumes:
      - application/json
      description: Create movie
      parameters:
      - description: Movie creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.MovieRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sqlc.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create movie
      tags:
      - movie
  /movie/{movie_id}:
    delete:
      consumes:
      - application/json
      description: Delete movie
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete movie
      tags:
      - movie
    get:
      consumes:
      - application/json
      description: Get movie by id
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.GetMovieByIDRow'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Show movie
      tags:
      - movie
    patch:
      consumes:
      - application/json
      description: Update movie
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      - description: Movie creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.MovieRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update movie
      tags:
      - movie
...
  /user/{user_id}/favorite_movie:
    get:
      consumes:
      - application/json
      description: Get user's favorite movie list
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.FavoriteMovieListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get favorite movie list
      tags:
      - favorite_movie
  /user/{user_id}/favorite_movie/{movie_id}:
    delete:
      consumes:
      - application/json
      description: Delete favorite movie
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete favorite movie
      tags:
      - favorite_movie
    post:
      consumes:
      - application/json
      description: Add movie to user's favorite
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.FavoriteMovie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add favorite movie
      tags:
      - favorite_movie
...
swagger: "2.0"